adding roles to the controllers

+++ src/main/java/org/nextprot/api/user/service/impl/UserQueryServiceImpl.java
@@ -6,7 +6,7 @@ import org.nextprot.api.commons.exception.NotAuthorizedException;
 import org.nextprot.api.user.dao.UserQueryDao;
 import org.nextprot.api.user.domain.UserQuery;
 import org.nextprot.api.user.domain.UserResource;
-import org.nextprot.api.user.security.NPContext;
+import org.nextprot.api.user.security.NPSecurityContext;
 import org.nextprot.api.user.service.UserQueryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
@@ -35,7 +35,7 @@ public class UserQueryServiceImpl implements UserQueryService {
 
 	@Override
 	public UserQuery createUserQuery(UserQuery userQuery) {
-		NPContext.isUserAuthorized(userQuery);
+		NPSecurityContext.checkUserAuthorization(userQuery);
 		userQuery.checkValid();
 		long id = userQueryDao.saveUserQuery(userQuery);
 		userQuery.setUserQueryId(id);
@@ -44,7 +44,7 @@ public class UserQueryServiceImpl implements UserQueryService {
 
 	@Override
 	public UserQuery updateUserQuery(UserQuery userQuery) {
-		NPContext.isUserAuthorized(userQuery);
+		NPSecurityContext.checkUserAuthorization(userQuery);
 		userQuery.checkValid();
 		userQueryDao.updateUserQuery(userQuery);
 		return userQuery;
@@ -52,7 +52,7 @@ public class UserQueryServiceImpl implements UserQueryService {
 
 	@Override
 	public void deleteUserQuery(UserQuery userQuery) {
-		NPContext.isUserAuthorized(userQuery);
+		NPSecurityContext.checkUserAuthorization(userQuery);
 		userQueryDao.deleteUserQuery(userQuery.getUserQueryId());
 	}
